refactor: Adjust namespaces
