+++ src/API/Controllers/MenuItemController.cs
@@ -1,5 +1,4 @@
-using Application.DTOs;
-using Application.Errors;
+using Application.DTOs.Request;
 using Application.Interfaces;
 using Microsoft.AspNetCore.Mvc;
