+++ src/Application/DTOs/Request/CreateMenuItemDTO.cs
@@ -1,8 +1,6 @@
-
-
 using System.ComponentModel.DataAnnotations;
 
-namespace Application.DTOs
+namespace Application.DTOs.Request
 {
     public class CreateMenuItemDTO
     {
