+++ src/Application/DTOs/Request/CreateOrderDTO.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace Application.DTOs
+namespace Application.DTOs.Request
 {
     public class CreateOrderDTO
     {
