+++ src/Application/DTOs/Request/UpdateOrderDTO.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel.DataAnnotations;
 using Domain.Enums;
 
-namespace Application.DTOs
+namespace Application.DTOs.Request
 {
     public class UpdateOrderDTO
     {
