+++ src/Application/DTOs/Response/MenuItemResponseDTO.cs
@@ -1,8 +1,4 @@
-
-
-using System.ComponentModel.DataAnnotations;
-
-namespace Application.DTOs
+namespace Application.DTOs.Response
 {
     public class MenuItemResponseDTO
     {
