+++ src/Application/DTOs/Response/OrderResponseDTO.cs
@@ -1,4 +1,4 @@
-namespace Application.DTOs
+namespace Application.DTOs.Response
 {
     public class OrderResponseDTO
     {
