+++ src/Application/Interfaces/ICustomerService.cs
@@ -1,4 +1,4 @@
-using Application.DTOs;
+using Application.DTOs.Response;
 
 namespace Application.Interfaces
 {
