+++ src/Application/Interfaces/IOrderService.cs
@@ -1,4 +1,5 @@
-using Application.DTOs;
+using Application.DTOs.Request;
+using Application.DTOs.Response;
 
 namespace Application.Interfaces
 {
