+++ src/Application/Services/MenuItemService.cs
@@ -1,9 +1,10 @@
 using Application.DTOs;
+using Application.DTOs.Request;
+using Application.DTOs.Response;
 using Application.Errors;
 using Application.Interfaces;
 using Application.Validators;
 using Domain.Entities;
-using Infrastructure.Persistence;
 using Infrastructure.Repositories.Interfaces;
 
 namespace Application.Services
